collection of fixes and changes to WC (#300)

+++ app/src/main/java/com/concordium/wallet/data/model/TransferCost.kt
@@ -1,9 +1,7 @@
 package com.concordium.wallet.data.model
 
-import java.math.BigInteger
-
 data class TransferCost(
-    val energy: BigInteger,
+    val energy: Long,
     val cost: String,
     val success: Boolean?
 )
